refactor(auditor): split rules into smaller modules
